moved themes to HeroDemo
