+++ components/Hero/Themes/NTO/Inventory.jsx
