+++ components/Hero/index.js
@@ -1,2 +1,3 @@
 export { Hero } from './Hero';
 export { HeroDemo } from './HeroDemo';
+export { Themes, ThemeSelector, Superstore, NTO, Ebikes } from './Themes';
