--- components/Themes/Superstore/Overview.jsx
@@ -1,32 +1,0 @@
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "components/ui";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "components/ui";
-
-import { TableauEmbed } from 'components';
-
-export const Overview = () => {
-  return (
-    <TabsContent value="overview" className="space-y-4">
-      <Card className='dark:bg-stone-900'>
-        <CardHeader>
-          <CardTitle>Overview</CardTitle>
-          <CardDescription>
-            Your personal digest of Superstore sales in North America
-          </CardDescription>
-        </CardHeader>
-        <CardContent className="flex items-center justify-center">
-          <TableauEmbed
-            src='https://prod-useast-b.online.tableau.com/t/embeddingplaybook/views/superstore/overview_800x800'
-            width={800}
-            height={1000}
-            hideTabs={true}
-            device='default'
-            toolbar='hidden'
-            layouts = {{
-              '*': { 'device': 'default', 'width': 800, 'height': 1100 }
-            }}
-          />
-        </CardContent>
-      </Card>
-    </TabsContent>
-  )
-}
